plugins: (spotifytoyoutube) Fix regex parsing issue
The character `$` in a song title was causing issues.

However, using `regexp.QuoteMeta` on its own doesn't solve the issue, because it also turns `|` into `\|`, when we *do* want to match a raw `|` instead, so we have to update our regex accordingly. This hopefully shouldn't break, as `|` was the only regex meta character that we were raw-matching.